[DEV] change uint8_t to uin16_t for startbit
        * needed to handle frame bigger than 8 bytes
          such as CANFD frames (64 bytes max)

+++ include/libwecan.h
@@ -12,15 +12,15 @@
 #define INTEL     4
 #define MOTOROLA  5
 
-void compute_indexes(const uint8_t startbit, const uint8_t length, uint8_t endianness, 
+void compute_indexes(const uint16_t startbit, const uint8_t length, uint8_t endianness, 
                 uint8_t *offset_lsb, uint8_t *byte_index_lsb, uint8_t *byte_index_msb)
 {
-        uint8_t msb = 0; 
-        uint8_t lsb = 0;
+        uint16_t msb = 0; 
+        uint16_t lsb = 0;
 
         if (endianness == MOTOROLA) {
                 msb =  startbit;
-                uint8_t current_bit_nr = startbit;
+                uint16_t current_bit_nr = startbit;
 
                 int i;
                 for (i = 0; i < length - 1; i++) {
@@ -77,7 +77,7 @@ uint64_t frame_to_local(const uint8_t *frame, uint8_t byte_index_lsb,
         return target;
 }
 
-uint64_t extract(const uint8_t *frame, const uint8_t startbit, 
+uint64_t extract(const uint8_t *frame, const uint16_t startbit, 
                 const uint8_t length, uint8_t signedness, uint8_t endianness)
 {
         /* 
@@ -118,7 +118,7 @@ uint64_t extract(const uint8_t *frame, const uint8_t startbit,
         return target;
 }
 
-void insert(uint8_t *frame, uint8_t startbit, uint8_t length,uint64_t value, 
+void insert(uint8_t *frame, uint16_t startbit, uint8_t length, uint64_t value, 
                 uint8_t endianness)
 {
         uint8_t byte_index_msb = 0;
@@ -162,56 +162,56 @@ void insert(uint8_t *frame, uint8_t startbit, uint8_t length,uint64_t value,
         }
 }
 
-void encode_uint64_t(uint8_t *frame, uint64_t physical_value, uint8_t startbit,
+void encode_uint64_t(uint8_t *frame, uint64_t physical_value, uint16_t startbit,
                 uint8_t length, uint8_t endianness, double factor, double offset)
 {
         uint64_t can_value = (physical_value - offset) / factor;
         insert(frame, startbit, length, can_value, endianness);
 }
 
-void encode_int64_t(uint8_t *frame, int64_t physical_value, uint8_t startbit,
+void encode_int64_t(uint8_t *frame, int64_t physical_value, uint16_t startbit,
                 uint8_t length, uint8_t endianness, double factor, double offset)
 {
         uint64_t can_value = (physical_value - offset) / factor;
         insert(frame, startbit, length, can_value, endianness);
 }
 
-void encode_double(uint8_t *frame, double physical_value, uint8_t startbit,
+void encode_double(uint8_t *frame, double physical_value, uint16_t startbit,
                 uint8_t length, uint8_t endianness, double factor, double offset)
 {
         uint64_t can_value = (uint64_t)((physical_value - offset) / factor);
         insert(frame, startbit, length, can_value, endianness);
 }
 
-void encode_float(uint8_t *frame, float physical_value, uint8_t startbit,
+void encode_float(uint8_t *frame, float physical_value, uint16_t startbit,
                 uint8_t length, uint8_t endianness, double factor, double offset)
 {
         uint64_t can_value = (uint64_t)((physical_value - offset) / factor);
         insert(frame, startbit, length, can_value, endianness);
 }
 
-uint64_t decode_uint64_t(uint8_t *frame, uint8_t startbit, uint8_t length, 
+uint64_t decode_uint64_t(uint8_t *frame, uint16_t startbit, uint8_t length, 
                 uint8_t endianness, double factor, double offset)
 {
         uint64_t can_value = extract(frame, startbit, length, UNSIGNED, endianness);
         return (can_value * factor) + offset;
 }
 
-int64_t decode_int64_t(uint8_t *frame, uint8_t startbit, uint8_t length, 
+int64_t decode_int64_t(uint8_t *frame, uint16_t startbit, uint8_t length, 
                 uint8_t endianness, double factor, double offset)
 {
         int64_t can_value = (int64_t)extract(frame, startbit, length, SIGNED, endianness);
         return (can_value * factor) + offset;
 }
 
-double decode_double(uint8_t *frame, uint8_t startbit, uint8_t length, 
+double decode_double(uint8_t *frame, uint16_t startbit, uint8_t length, 
                 uint8_t endianness, double factor, double offset)
 {
         int64_t can_value = (int64_t)extract(frame, startbit, length, SIGNED, endianness);
         return (can_value * factor) + offset;
 }
 
-float decode_float(uint8_t *frame, uint8_t startbit, uint8_t length, 
+float decode_float(uint8_t *frame, uint16_t startbit, uint8_t length, 
                 uint8_t endianness, double factor, double offset)
 {
         int64_t can_value = (int64_t)extract(frame, startbit, length, SIGNED, endianness);
